fix: 29669359, 6614221, 69890967, 85121942, 88071625, 98645731
29669359 No.61 ヴォルカザウルス
Now No.61 ヴォルカザウルス can destroy a monster which is changed into face down in chain (By Konami database FAQ), and the attack caluculation is changed  into GetTextAttack().

6614221 霧の王
Now the attack caluculation is changed  into GetTextAttack().

69890967  幻魔皇ラビエル
Add the check for monster with text attack '?'.

85121942  CNo.105 BK 彗星のカエストス
Now damage when battle destroying a monster is calculated by GetTextAttack().

88071625 Ｔｈｅ　ｔｙｒａｎｔ　ＮＥＰＴＵＮＥ
Now the attck and deffense calculation is using GetTextAttack(), and the
summation is changed into the way similar to 6614221.

98645731  強欲で謙虚な壺
Now the card ID in Flag Effect is the correct ID.