fix
440556 バハムート·シャーク
The player who activates sp_summon effect will sp_summon the water xyz
monster, even if the control is changed to opponent.
(According to Konami database)

69042950 バグ·ロード
Revert to  previous version.
(This card can be activated only if both player control a monster with
level<=4)
The sp_summon of opponent is selectable now.

75797046 フォトン・アレキサンドラ・クイーン
The previous ID is wrong, and it should be 75797046.