fix
56840427 CNo.39 希望皇ホープレイ

http://www.db.yugioh-card.com/yugiohdb/faq_search.action?ope=5&fid=12335&keyword=&tag=-1
The ATK increasing effect cannot be disabled after applying to this
card.

68836428 トライエッジ・リヴァイア
EFFECT_FLAG_CANNOT_DISABLE is added to the ATK decreasing effect, as all
similar effects that can target the card itself.
Now the negate effect can work on disabled monster.

1639384 神竜騎士フェルグラント
Now the negate effect can work on disabled monster.

78474168 ブレイクスルー·スキル
The condition is changed into the form similar to エフェクト・ヴェーラー.

82732705 スキルドレイン
Now it can negate the effect of dual monster and Noble Knight properly.