fix: multiple target
c21007444 諸刃の活人剣術
http://www.db.yugioh-card.com/yugiohdb/faq_search.action?ope=5&fid=12356&keyword=&tag=-1
If the location count of MZONE is less than 2, the player can choose 1 card form the 2 target.
The destroy is effect is changed into the form similar to c27174286 異次元からの帰還

c66957584 インフェルニティガン
If the location count of M_ZONE is less than 2, it will not special summon any monster.

c86197239 インフェルニティ・ミラージュ
fix card name

c22842214 シー・ランサー
http://www.db.yugioh-card.com/yugiohdb/faq_search.action?ope=5&fid=6643&keyword=&tag=-1
1. If it is not on the field when its effect resolves, the target cards should be sent ot grave.
2. If the location of SZONE is not enough, the player can choose appropriate number of card form those targets.

c92901944 スクラップ・サーチャー
The sp_summon effect cannot be activated if スクラップ・サーチャー itself is in the group of destroyed Scrap monsters.