fix
c96142517 RUM－千死蛮巧
The sp_summon is not treated as XYZ summon.

シャドール
C1 ブラック・ホール
C2 堕ち影の蠢き
This will make it possible for the player to activate any one of the 2 effect at the same time, and the opponent may not know which effect is being activated.
Now シャドール monster effects will show the effect number while being activated.