fix: attacker
c21501505 暗遷士 カンゴルゴーム
http://www.db.yugioh-card.com/yugiohdb/faq_search.action?ope=4&cid=11083
EFFECT_FLAG_CARD_TARGET is added since it is an target effect.

c64726269 光天使スケール
HintSelection is added to show which card is returned to deck.

c60080151 好敵手の記憶
http://www.db.yugioh-card.com/yugiohdb/faq_search.action?ope=4&cid=10297
It is not an target effect, so the attacker should be determined while resolving the effect.
The condition of Attacking Moster is changed into the form similar to c70342110 次元幽閉.

c79997591 ドゥーブルパッセ
http://www.db.yugioh-card.com/yugiohdb/faq_search.action?ope=5&fid=13260&keyword=&tag=-1
http://www.db.yugioh-card.com/yugiohdb/faq_search.action?ope=5&fid=13246&keyword=&tag=-1
http://www.db.yugioh-card.com/yugiohdb/faq_search.action?ope=5&fid=13261&keyword=&tag=-1
The condition of Attacking Moster and Attacked Target is changed into the form similar to c70342110 次元幽閉.