fix
If the "is unaffected by the effects of other Spell/Trap Cards." effect
of "Forbidden Lance" is applied to a monster, "Skill Prisoner" can still
be activated by targeting that monster. Even in this case, the effect of
"Skill Prisoner" is applied normally. (Since the effect of "Skill
Prisoner" has been applied, any monster effects that target that monster
will be negated.)[2]

http://www.db.yugioh-card.com/yugiohdb/faq_search.action?ope=5&fid=12980&request_locale=ja