fix
c65305468 FNo.0 未来皇ホープ
The rank 1 statemnet is not a part of its effect, so it is a rank 1 xyz monster.
It should be rank 1 in cards.cdb since its original rank is 1, and this effect is not needed.

c66127916 融合準備
Now it can send the monster to hand and get 融合 from grave.

c67249508 竜星の凶暴化
Now the atk changing effect has correct target.

c31077447 びっくり箱
Now the atk decreasing effect is permanant.