Update c7165085.lua
Semi Fix: This fix enables Bait Doll to activate the a Trap which has only just been set on your opponents side of the filed if they control Temple of the Kings. More work is needed to fix it so it will also work if the turn player uses it on their own just set Trap while they Control Temple of the Kings and it should only work once per turn like this due to the new Temple of the Kings errata.