Add wildcard support in deck edit (#6)
* Add wildcard support in deck edit

If a search term contains `*`, the game will return results for which any text is in place of `*`.

* Improve wildcard declaration

ficks

* Improve wildcard handling

Can now detect multiple instances of character immediately after wildcard.

For example, Dinomist Plesios has "Dinomist" in both its Pendulum Effect and Monster effect. Before, if you searched text that matched before `"*` in the Pendulum text, and after `*"` in the monster text, it would terminate on the `"` in the Pendulum text and not match, but now it keeps trying when it finds `"` again in the monster text and successfully matches.