Fixed resources cleaning when closing the client during a duel for non windows
The client would get stuck because of mutexes being locked and then non unloched because the main class got deleted, added additinal mutex check when analyzing, to make sure that the current message finished analyzing before attempting to stop the client, moved singleSignal to the SingleMode class, as it's more appropriate. Call clear texture when closing so the threads are stopped.