Update whitelist handling
Don't automatically allow cards if the card they're aliased to is present in the whitelist, unless their codes are in the artwork offset range. This solves the issue with random cards being allowed in a card pool by simply being aliased to a card in it. Also cleaned up some code and added some helper functions.